feature: begin work on building Express backend to query database, refactor front-end
