+++ index.html
@@ -7,6 +7,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="stylesheet" type="text/css" href="searchy.css">
+  <script type="text/javascript" src="script.js"></script>
 </head>
 
 <body>
@@ -34,9 +35,6 @@
   </div>
 
   <button data-js="export-button">Export List</button>
-
-  <script src="https://cdn.rawgit.com/Keyang/node-csvtojson/d41f44aa/browser/csvtojson.min.js"></script>
-  <script type="text/javascript" src="searchy.js"></script>
 </body>
 
 </html>
