+++ package.json
@@ -12,16 +12,18 @@
   "license": "ISC",
   "dependencies": {
     "csvtojson": "^2.0.10",
+    "express": "^4.17.1",
     "lodash": "^4.17.15",
+    "pg": "^8.5.1",
     "sass": "^1.26.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
     "@babel/preset-env": "^7.9.5",
     "babel-plugin-lodash": "^3.3.4",
-    "browserify": "^16.5.0",
     "babel-preset-es2015": "^6.6.0",
     "babelify": "^7.3.0",
+    "browserify": "^16.5.0",
     "uglifyify": "^3.0.4"
   }
 }
